requirements.txt
confluent-kafka==2.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
flask==3.0.0